refactor: 拆分 evidence 模块，删除 HearingNote，用 proxy model 统一 admin 入口
- 删除 HearingNote 模型、admin 和相关测试
- 删除 documents/admin/evidence/ 死代码（旧版 evidence admin 副本）
- 创建 EvidenceListProxy/EvidenceItemProxy（app_label='evidence'）
- EvidenceList/EvidenceItem admin 从 /admin/documents/ 移到 /admin/evidence/
- 更新 hearing_mode.py 移除 HearingNote 引用
- 更新测试文件中的 import 路径
