--- backend/apps/documents/admin/evidence/__init__.py
@@ -1,4 +1,0 @@
-from .forms import EvidenceListForm
-from .inlines import EvidenceItemInline
-
-__all__ = ["EvidenceItemInline", "EvidenceListForm"]
