--- backend/apps/documents/admin/evidence/forms.py
@@ -1,24 +1,0 @@
-"""Module for forms."""
-
-from __future__ import annotations
-
-from typing import Any
-
-from django import forms
-
-from apps.documents.models import EvidenceList
-
-
-class EvidenceListForm(forms.ModelForm[EvidenceList]):  # pragma: no cover
-    class Meta:  # pragma: no cover
-        model = EvidenceList
-        fields: str = "__all__"
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:  # pragma: no cover
-        super().__init__(*args, **kwargs)
-
-        if "list_type" in self.fields:
-            self.fields["list_type"].disabled = True
-
-
-__all__: list[str] = ["EvidenceListForm"]
