--- backend/apps/documents/admin/evidence/inlines.py
@@ -1,52 +1,0 @@
-"""Module for inlines."""
-
-from __future__ import annotations
-
-from typing import Any, ClassVar
-
-from django.contrib import admin
-from django.db.models import QuerySet
-from django.http import HttpRequest
-
-from apps.documents.models import EvidenceItem
-
-
-class EvidenceItemInline(admin.TabularInline[EvidenceItem, EvidenceItem]):  # pragma: no cover
-    model = EvidenceItem
-    extra: int = 1
-    fields: ClassVar[tuple[Any, ...]] = (
-        "global_order_display",
-        "name",
-        "purpose",
-        "file",
-        "page_count",
-        "page_range_display",
-    )
-    readonly_fields: ClassVar[tuple[Any, ...]] = ("global_order_display", "page_count", "page_range_display")
-    ordering: ClassVar[list[str]] = ["order"]  # type: ignore[assignment]
-
-    def get_queryset(self, request: HttpRequest) -> QuerySet[EvidenceItem, EvidenceItem]:  # pragma: no cover
-        qs: QuerySet[EvidenceItem, EvidenceItem] = super().get_queryset(request)
-        return qs
-
-    @admin.display(description="序号")
-    def global_order_display(self, obj: EvidenceItem) -> Any:  # pragma: no cover
-        if not obj.pk:
-            return "-"
-
-        evidence_list = obj.evidence_list
-        global_order = evidence_list.start_order + obj.order - 1
-        return global_order
-
-    @admin.display(description="页码范围")
-    def page_range_display(self, obj: EvidenceItem) -> Any:  # pragma: no cover
-        if obj.pk:
-            return obj.page_range_display
-        return "-"
-
-    class Media:  # pragma: no cover
-        css: ClassVar[dict[str, tuple[str, ...]]] = {"all": ("documents/css/evidence_inline.css",)}
-        js: ClassVar[tuple[str, ...]] = ("documents/js/evidence_sortable.js",)
-
-
-__all__: list[str] = ["EvidenceItemInline"]
