--- backend/apps/documents/admin/evidence/mixins/__init__.py
@@ -1,9 +1,0 @@
-from .actions import EvidenceListAdminActionsMixin
-from .save import EvidenceListAdminSaveMixin
-from .views import EvidenceListAdminViewsMixin
-
-__all__ = [
-    "EvidenceListAdminActionsMixin",
-    "EvidenceListAdminSaveMixin",
-    "EvidenceListAdminViewsMixin",
-]
