+++ backend/apps/evidence/admin/__init__.py
@@ -1,4 +1,3 @@
 from .evidence_admin import EvidenceListAdmin
 from .evidence_item_admin import EvidenceItemAdmin
 from .group_admin import EvidenceGroupAdmin
-from .hearing_note_admin import HearingNoteAdmin
