+++ backend/tests/ci/unit/evidence/test_evidence_model.py
@@ -1,4 +1,4 @@
-"""Evidence Model 测试 - EvidenceList, EvidenceItem, EvidenceGroup, HearingNote"""
+"""Evidence Model 测试 - EvidenceList, EvidenceItem, EvidenceGroup"""
 
 from __future__ import annotations
 
@@ -10,7 +10,6 @@ from apps.evidence.models import (
     EvidenceList,
     EvidenceItem,
     EvidenceGroup,
-    HearingNote,
     ListType,
     EvidenceType,
     EvidenceDirection,
@@ -129,25 +128,3 @@ class TestEvidenceGroupModel:
         group = EvidenceGroup.objects.create(case=case, name="分组1", sort_order=1)
         group.items.add(item1, item2)
         assert group.items.count() == 2
-
-
-@pytest.mark.django_db
-class TestHearingNoteModel:
-    """HearingNote 模型测试"""
-
-    def test_create_note(self) -> None:
-        """创建庭审笔记"""
-        case = _create_case()
-        note = HearingNote.objects.create(case=case, content="庭审笔记内容")
-        assert note.content == "庭审笔记内容"
-
-    def test_note_with_evidence_items(self) -> None:
-        """庭审笔记关联证据"""
-        case = _create_case()
-        elist = EvidenceList.objects.create(
-            case=case, title="笔记测试清单", list_type="previous", order=1
-        )
-        item = EvidenceItem.objects.create(evidence_list=elist, order=1, name="笔记证据")
-        note = HearingNote.objects.create(case=case, content="关联证据笔记")
-        note.evidence_items.add(item)
-        assert note.evidence_items.count() == 1
